Create stringInt2BaseN
Take a string which holds an integer in base 10. And also as a parameter of an integer representing the base of your choice. 

Then convert that into an integer that represents the new number in the base of your choice...while removing any leading zeroes etc. 

(NOTE  -> This only works on 32 bit integer manipulation :D ) -Tyler